add pasta script para limpar o db assim q possivel , mudanca no console error, add ong nao encontrada e logica para o prazo necessidade , nao permite datas no passado ,novas pastas mais seletivas de integracao.
