+++ tests/integracao/doacoes-put.test.js
@@ -1,0 +1,232 @@
+const request = require('supertest');
+const jwt = require('jsonwebtoken');
+const app = require('../../index'); 
+describe('Doações - PUT (atualização)', () => {
+    let token;
+    let tokenOutraOng;
+    let doacaoId;
+
+    beforeAll(async () => {
+        // Token da ONG principal
+        const login = await request(app)
+            .post('/auth/login')
+            .send({ email: process.env.TEST_EMAIL, password: process.env.TEST_PASSWORD });
+        token = login.body.token;
+        console.log('TESTE: id_ong do token principal:', jwt.decode(token).id_ong);
+
+        // Simular token de outra ONG (para teste 403)
+        tokenOutraOng = jwt.sign(
+            { id_ong: 999, email: 'outra@ong.com' }, 
+            process.env.JWT_SECRET, 
+            { expiresIn: '1h' }
+        );
+        console.log('TESTE: id_ong do tokenOutraOng:', jwt.decode(tokenOutraOng).id_ong);
+
+        // Criar uma doação para testar a atualização
+        const novaDoacao = {
+            titulo: 'Doação para atualizar',
+            descricao: 'Descrição original',
+            tipo_item: 'Utensílios Gerais',
+            quantidade: 5,
+            prazo_necessidade: '2023-12-31',
+            url_imagem: 'https://exemplo.com/imagem.jpg',
+            urgencia: 'MEDIA',
+            whatsapp: '11999999999',
+            email: 'teste@exemplo.com'
+        };
+        const response = await request(app)
+            .post('/doacoes')
+            .set('Authorization', `Bearer ${token}`)
+            .send(novaDoacao);
+        doacaoId = response.body.id_produto;
+        console.log('ONG ID da doação criada:', response.body.ong_id);
+        console.log('ONG ID do token:', jwt.decode(token).id_ong);
+    });
+
+    // operação correta
+    it('deve atualizar doação com token da ONG dona', async () => {
+        const dadosAtualizados = {
+            titulo: 'Doação Atualizada',
+            descricao: 'Nova descrição',
+            tipo_item: 'Utensílios Gerais',
+            quantidade: 10,
+            prazo_necessidade: '2023-12-31',
+            url_imagem: 'https://exemplo.com/imagem.jpg',
+            urgencia: 'MEDIA',
+            whatsapp: '11999999999',
+            email: 'teste@exemplo.com'
+        };
+
+        const response = await request(app)
+            .put(`/doacoes/${doacaoId}`)
+            .set('Authorization', `Bearer ${token}`)
+            .send(dadosAtualizados);
+
+        console.log('Status retornado:', response.statusCode);
+        console.log('Body retornado:', response.body);
+
+        expect(response.statusCode).toBe(200);
+        expect(response.body.titulo).toBe(dadosAtualizados.titulo);
+    });
+
+    // teste para erro 401
+    it('deve retornar erro 401 ao tentar atualizar sem token', async () => {
+        const response = await request(app)
+            .put(`/doacoes/${doacaoId}`)
+            .send({ titulo: 'Tentativa sem token' });
+
+        expect(response.statusCode).toBe(401);
+        expect(response.body).toHaveProperty('message');
+    });
+
+    // teste para erro 403
+    it('deve retornar erro 403 ao tentar atualizar doação de outra ONG', async () => {
+        const dadosAtualizados = {
+            titulo: 'Doação Atualizada',
+            descricao: 'Nova descrição',
+            tipo_item: 'Utensílios Gerais',
+            quantidade: 10,
+            prazo_necessidade: '2023-12-31',
+            url_imagem: 'https://exemplo.com/imagem.jpg',
+            urgencia: 'MEDIA',
+            whatsapp: '11999999999',
+            email: 'teste@exemplo.com'
+        };
+        const response = await request(app)
+            .put(`/doacoes/${doacaoId}`)
+            .set('Authorization', `Bearer ${tokenOutraOng}`)
+            .send(dadosAtualizados);
+        expect(response.statusCode).toBe(403);
+        expect(response.body).toHaveProperty('message');
+        expect(response.body.message).toContain('Você não tem permissão para modificar esta doação');
+    });
+
+    // teste para erro 404 ( id inexistente) 
+    it('deve retornar erro 404 para ID inexistente', async () => {
+        const dadosAtualizados = {
+            titulo: 'Doação Atualizada',
+            descricao: 'Nova descrição',
+            tipo_item: 'Utensílios Gerais',
+            quantidade: 10,
+            prazo_necessidade: '2023-12-31',
+            url_imagem: 'https://exemplo.com/imagem.jpg',
+            urgencia: 'MEDIA',
+            whatsapp: '11999999999',
+            email: 'teste@exemplo.com'
+        };
+        const response = await request(app)
+            .put('/doacoes/99999')
+            .set('Authorization', `Bearer ${token}`)
+            .send(dadosAtualizados);
+        expect(response.statusCode).toBe(404);
+        expect(response.body).toHaveProperty('message');
+    });
+
+    // teste para erro 400 - campos obrigatórios PUT
+    const camposObrigatorios = [
+      'titulo', 'descricao', 'tipo_item', 'prazo_necessidade', 'url_imagem', 'urgencia', 'whatsapp', 'email'
+    ];
+    camposObrigatorios.forEach(campo => {
+      it(`deve retornar erro 400 no PUT se campo obrigatório '${campo}' estiver ausente ou vazio`, async () => {
+        const doacaoValida = {
+          titulo: 'Teste',
+          descricao: 'desc',
+          tipo_item: 'Roupas e Calçados',
+          quantidade: 1,
+          prazo_necessidade: '2023-12-31',
+          url_imagem: 'https://exemplo.com/imagem.jpg',
+          urgencia: 'MEDIA',
+          whatsapp: '11999999999',
+          email: 'teste@exemplo.com'
+        };
+        doacaoValida[campo] = '';
+        const response = await request(app)
+          .put(`/doacoes/${doacaoId}`)
+          .set('Authorization', `Bearer ${token}`)
+          .send(doacaoValida);
+        expect(response.statusCode).toBe(400);
+        expect(response.body).toHaveProperty('message');
+        expect(response.body.message).toContain(campo);
+      });
+    });
+    it('deve retornar erro 400 para quantidade negativa', async () => {
+      const dadosInvalidos = {
+        titulo: 'Teste',
+        descricao: 'desc',
+        tipo_item: 'Roupas e Calçados',
+        quantidade: -5,
+        prazo_necessidade: '2023-12-31',
+        url_imagem: 'https://exemplo.com/imagem.jpg',
+        urgencia: 'MEDIA',
+        whatsapp: '11999999999',
+        email: 'teste@exemplo.com'
+      };
+      const response = await request(app)
+        .put(`/doacoes/${doacaoId}`)
+        .set('Authorization', `Bearer ${token}`)
+        .send(dadosInvalidos);
+      expect(response.statusCode).toBe(400);
+      expect(response.body).toHaveProperty('message');
+      expect(response.body.message).toContain('Quantidade');
+    });
+    it('deve retornar erro 400 para email inválido no PUT', async () => {
+      const doacao = {
+        titulo: 'Teste', descricao: 'desc', tipo_item: 'Roupas e Calçados', quantidade: 1,
+        prazo_necessidade: '2023-12-31', url_imagem: 'https://exemplo.com/imagem.jpg', urgencia: 'MEDIA', whatsapp: '11999999999', email: 'emailinvalido'
+      };
+      const response = await request(app)
+        .put(`/doacoes/${doacaoId}`)
+        .set('Authorization', `Bearer ${token}`)
+        .send(doacao);
+      expect(response.statusCode).toBe(400);
+      expect(response.body.message).toContain('Email');
+    });
+
+    it('deve retornar erro 400 para whatsapp inválido no PUT', async () => {
+      const doacao = {
+        titulo: 'Teste', descricao: 'desc', tipo_item: 'Roupas e Calçados', quantidade: 1,
+        prazo_necessidade: '2023-12-31', url_imagem: 'https://exemplo.com/imagem.jpg', urgencia: 'MEDIA', whatsapp: 'abc123', email: 'teste@exemplo.com'
+      };
+      const response = await request(app)
+        .put(`/doacoes/${doacaoId}`)
+        .set('Authorization', `Bearer ${token}`)
+        .send(doacao);
+      expect(response.statusCode).toBe(400);
+      expect(response.body.message).toContain('Whatsapp');
+    });
+
+    it('deve retornar erro 400 para url_imagem inválida no PUT', async () => {
+      const doacao = {
+        titulo: 'Teste', descricao: 'desc', tipo_item: 'Roupas e Calçados', quantidade: 1,
+        prazo_necessidade: '2023-12-31', url_imagem: 'imagem_invalida', urgencia: 'MEDIA', whatsapp: '11999999999', email: 'teste@exemplo.com'
+      };
+      const response = await request(app)
+        .put(`/doacoes/${doacaoId}`)
+        .set('Authorization', `Bearer ${token}`)
+        .send(doacao);
+      expect(response.statusCode).toBe(400);
+      expect(response.body.message).toContain('url_imagem');
+    });
+
+    it('deve retornar erro 400 para urgencia inválida no PUT', async () => {
+      const doacao = {
+        titulo: 'Teste', descricao: 'desc', tipo_item: 'Roupas e Calçados', quantidade: 1,
+        prazo_necessidade: '2023-12-31', url_imagem: 'https://exemplo.com/imagem.jpg', urgencia: 'URGENTE', whatsapp: '11999999999', email: 'teste@exemplo.com'
+      };
+      const response = await request(app)
+        .put(`/doacoes/${doacaoId}`)
+        .set('Authorization', `Bearer ${token}`)
+        .send(doacao);
+      expect(response.statusCode).toBe(400);
+      expect(response.body.message).toContain('Urgência');
+    });
+
+    it('deve retornar erro 400 se body estiver vazio no PUT', async () => {
+      const response = await request(app)
+        .put(`/doacoes/${doacaoId}`)
+        .set('Authorization', `Bearer ${token}`)
+        .send({});
+      expect(response.statusCode).toBe(400);
+      expect(response.body).toHaveProperty('message');
+    });
+});
